Add DAT files marked with their version field.
These are collected from this repo as well as found in other repos across
github from people who have synced their character data to github in ages
past (as well as my own gearset file). This doesn't represent a complete
view of all the DAT versions released, but should be enough to deal with
compatibility testing.